% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_tsNEM_dataset}
\alias{generate_tsNEM_dataset}
\title{Generate an effect profile from a static NEM.}
\usage{
generate_tsNEM_dataset(nemObject, perturb = NULL)
}
\arguments{
\item{nemObject}{A `nem` object as generated by `sample_random_NEM_and_data`.}

\item{perturb}{The name of an S-node to perturb.}
}
\description{
Generate an effect profile from a static NEM.
}
