% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsNEM.R
\name{rowMaker}
\alias{rowMaker}
\title{Internal function for making a row in the canonical ILP constraint matrix.}
\usage{
rowMaker(edgeCoordinates, dimension)
}
\arguments{
\item{edgeCoordinates}{Node indices of an edge in the adjacency matrix of the
transitively reduced signalling graph.}

\item{dimension}{The total number of nodes in the signalling graph.}
}
\value{
A row of length \code{dimension} for the canonical ILP constraint
  matrix.
}
\description{
Internal function for making a row in the canonical ILP constraint matrix.
}
