% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsNEM.R
\name{runILP}
\alias{runILP}
\title{Internal function for running the ILP.}
\usage{
runILP(NamedEgeneProfile, nemObject, constraintMatrix)
}
\arguments{
\item{NamedEgeneProfile}{Effect profile as a named vector.}

\item{nemObject}{A \code{nem} object.}

\item{constraintMatrix}{The corresponding canonical ILP constraint matrix.}
}
\value{
The ILP solution of the signalling states of the nodes in the graph.
}
\description{
Internal function for running the ILP.
}
