% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sample_random_NEM_and_data}
\alias{sample_random_NEM_and_data}
\title{Generate a random NEM.}
\usage{
sample_random_NEM_and_data(n = 10, m = 1000)
}
\arguments{
\item{n}{The number of S-nodes}

\item{m}{The number of E-genes}
}
\value{
A `nem` object.
}
\description{
Generate a random NEM.
}
