% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsNEM.R
\name{tsNEM}
\alias{tsNEM}
\title{Map an observational E-gene profile onto a NEM.}
\usage{
tsNEM(nemObject, observationalLogDensities)
}
\arguments{
\item{nemObject}{A \code{nem} object.}

\item{observationalLogDensities}{A matrix (or vector) of effect profiles
recorded as log densities. The (row) names must correspond to the E-gene
names in \code{nemObject}. The column names should identify the
observational profiles, for instance with time points.}
}
\value{
A matrix of signalling states for the nodes (rows) in the various
  time points (columns).
}
\description{
Map an observational E-gene profile onto a NEM.
}
